# docs/openapi/v1-ai-chat.yaml
openapi: 3.0.3
info:
  title: AI 成长计划 API — AI 对话与通知
  description: |
    AI 智能对话、助手任务查询、站内通知及 WebSocket 推送约定。
    鉴权方式：Bearer JWT（`bearerAuth`）。
    人类可读说明见仓库 `md文档/HTTP接口-AI对话与通知.md`、`md文档/AI对话与助手任务-流程说明.md`。
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: 本地开发

tags:
  - name: AI 对话
    description: 智能对话与会话消息
  - name: 助手任务
    description: 用户助手待办（非成长计划 tasks）
  - name: 站内通知
    description: 任务提醒等通知
  - name: 移动推送
    description: uni-push 2.0 clientId（锁屏/杀进程系统通知）

security:
  - bearerAuth: []

paths:
  /api/v1/ai/chat:
    post:
      tags: [AI 对话]
      summary: 发送对话消息
      description: |
        用户发一轮消息；服务端可选执行数据规划与意图分析，调用大模型及任务工具，
        将本轮用户消息与助手最终回复落库，并可通过 WebSocket 推送 CHAT_REPLY。
      operationId: postAiChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiChatRequest'
            example:
              message: 帮我记一下周五前要交周报
              sessionId: 12
              provider: mimo
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiChatResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: AI 提供商未配置（FEATURE_UNAVAILABLE）

  /api/v1/ai/chat/sessions:
    get:
      tags: [AI 对话]
      summary: 分页列出当前用户的历史会话
      description: |
        返回当前用户全部 AI 对话会话摘要（含 sessionId），按 `updatedAt` 降序。
        仅 JWT 对应用户本人的会话；不暴露他人数据。
      operationId: listAiChatSessions
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从 0 起
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页条数，最大 100
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiChatSessionPageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/ai/chat/sessions/{sessionId}/messages:
    get:
      tags: [AI 对话]
      summary: 分页拉取会话历史消息
      description: 按时间升序分页返回已落库的 USER/ASSISTANT 消息；含「任务提醒」会话。须为当前用户所属会话。
      operationId: listAiChatMessages
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiChatMessageListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/users/me/tasks:
    get:
      tags: [助手任务]
      summary: 查询当前用户的助手任务清单
      operationId: listMyAssistantTasks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [OPEN, DONE, CANCELLED]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAssistantTaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/users/me/notifications/unread-count:
    get:
      tags: [站内通知]
      summary: 未读通知数量
      operationId: getNotificationUnreadCount
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/users/me/notifications:
    get:
      tags: [站内通知]
      summary: 通知列表
      operationId: listNotifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InAppNotificationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/users/me/notifications/{id}/read:
    patch:
      tags: [站内通知]
      summary: 标记单条通知已读
      operationId: markNotificationRead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 已标记
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/users/me/notifications/read-all:
    patch:
      tags: [站内通知]
      summary: 全部标记已读
      operationId: markAllNotificationsRead
      responses:
        '204':
          description: 已全部标记
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/users/me/push-tokens:
    put:
      tags: [移动推送]
      summary: 注册或更新 uni-push clientId
      description: uni.getPushClientId() 上报；用于系统通知栏（后台/杀进程）。须 JWT。
      operationId: upsertPushToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPushTokenRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushTokenRegisteredResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [移动推送]
      summary: 注销推送 token（登出）
      operationId: deletePushToken
      parameters:
        - name: deviceId
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
          description: 不传则删除该用户全部已注册设备
      responses:
        '204':
          description: 已注销
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录或注册返回的 accessToken

  schemas:
    AiChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          maxLength: 8000
          description: 用户本轮输入
        sessionId:
          type: integer
          format: int64
          nullable: true
          description: 继续已有会话；不传则新建
        provider:
          type: string
          description: mimo（默认）或 ollama

    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
        sessionId:
          type: integer
          format: int64
        provider:
          type: string
        model:
          type: string
        userMessageId:
          type: integer
          format: int64
        assistantMessageId:
          type: integer
          format: int64
        planProposal:
          nullable: true
          allOf:
            - $ref: 'v1-growth-plan-proposals.yaml#/components/schemas/AiChatPlanProposalHint'
          description: 本轮若已生成待确认成长计划草案则非空

    AiChatMessageItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          type: string
          enum: [USER, ASSISTANT]
        content:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    AiChatSessionItemResponse:
      type: object
      properties:
        sessionId:
          type: integer
          format: int64
        title:
          type: string
          nullable: true
        provider:
          type: string
        model:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    AiChatSessionPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AiChatSessionItemResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean

    AiChatMessageListResponse:
      type: object
      properties:
        sessionId:
          type: integer
          format: int64
        sessionTitle:
          type: string
          nullable: true
        count:
          type: integer
          description: 本页消息条数
        messages:
          type: array
          items:
            $ref: '#/components/schemas/AiChatMessageItemResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean

    UserAssistantTaskItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [OPEN, DONE, CANCELLED]
        dueDate:
          type: string
          format: date
          nullable: true
        reminderSentAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    UserAssistantTaskListResponse:
      type: object
      properties:
        count:
          type: integer
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/UserAssistantTaskItemResponse'

    UnreadCountResponse:
      type: object
      properties:
        count:
          type: integer
          format: int64

    InAppNotificationItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          example: TASK_DUE_REMINDER
        title:
          type: string
        body:
          type: string
        taskId:
          type: integer
          format: int64
          nullable: true
        sessionId:
          type: integer
          format: int64
          nullable: true
        messageId:
          type: integer
          format: int64
          nullable: true
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    InAppNotificationListResponse:
      type: object
      properties:
        count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/InAppNotificationItemResponse'

    UpsertPushTokenRequest:
      type: object
      required: [platform, token, deviceId]
      properties:
        platform:
          type: string
          enum: [ANDROID, IOS]
        token:
          type: string
          maxLength: 512
          description: uni-push push_clientid（CID）
        deviceId:
          type: string
          maxLength: 64
          description: 客户端稳定设备 ID

    PushTokenRegisteredResponse:
      type: object
      properties:
        deviceId:
          type: string
        platform:
          type: string

    ErrorBody:
      type: object
      properties:
        code:
          type: string
        message:
          type: string

  responses:
    Unauthorized:
      description: 未登录或令牌无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            code: UNAUTHORIZED
            message: 未授权
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            code: NOT_FOUND
            message: 对话会话不存在
    ValidationError:
      description: 请求体验证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            code: VALIDATION_ERROR
            message: message 不能为空

# WebSocket（非 OpenAPI 标准 path，仅作文档补充）
# ws://{host}/ws/v1/chat?token={accessToken}
# 推送 type: CHAT_REPLY | TASK_DUE_REMINDER，字段见 md文档/HTTP接口-AI对话与通知.md 第 5 章